CodecRegistry is now an interface, BasicCodecRegistry its implementation; it is no longer a singleton
